pyproject.toml
[tool.poetry]
name = "kazspell"
version = "0.1.0"
description = "Spell Corrector for kazakh texts"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
metaflow = "^2.12.28"
nltk = "^3.9.1"
ebooklib = "^0.18"
beautifulsoup4 = "^4.12.3"
torch = "2.0.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
kubernetes = "^32.0.0"
pynvml = "^12.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

